+++ example/error_handler_example.dart
@@ -5,18 +5,18 @@ import 'post.dart';
 
 /// first create [Dio] api call
 FutureResponse<Post> getPost() async {
-  final dio = Dio();
-
-  const path = "https://jsonplaceholder.typicode.com/posts/1";
-
-  final response = await dio.get(path);
+  final response = await Dio().get(
+    "https://jsonplaceholder.typicode.com/posts/1",
+  );
 
   return response.convert(Post.fromJson);
 }
 
 /// wrap the api call with [safeApiCall]
 void main() {
-  safeApiCall(getPost).listen((event) {
+  final errorHandler = ErrorHandler();
+
+  errorHandler.stream(getPost).listen((event) {
     event.when(
       idle: () {
         print("init");