+++ lib/error_handler.dart
@@ -1,17 +1,11 @@
-/// Error handler And api request state handling
+/// error handler for all http client in dart dio, chopper http and more
+///
+/// freezed based api
 library error_handler;
 
 export 'error_handler_base.dart';
-
-export 'src/http_response.dart';
-export 'src/network_exception.dart';
-export 'src/result_state.dart';
-export 'src/result_state_extension.dart';
-export 'src/logger.dart';
-export 'src/http_response.dart';
-export 'src/http_response_extension.dart';
-export 'error_handler_base.dart';
 export 'error_handler_direct.dart';
-
-export 'src/network_exception_delegate.dart' show NetworkExceptionDelegate;
-export 'src/network_exception_localizations_utils.dart';
+export 'src/http_response/http_response.dart';
+export 'src/logger.dart';
+export 'src/network_exception/network_exception.dart';
+export 'src/result_state/result_state.dart';