+++ lib/src/http_response/http_response_base.dart
@@ -1,16 +1,25 @@
+import 'package:error_handler/error_handler.dart';
+
 /// future value of [HttpResponse]
 typedef FutureResponse<T> = Future<HttpResponse<T, dynamic>>;
 
-/// ResultState callback value
-/// T: formatted [data] type
-/// RT: [response] type
+/// response wrapper
+///
+/// [response] will be passed to [ErrorHandler]
+///
+/// - [data] formatted data type of [T]
+/// - [response] response value to used it [ResultState.error] of type [RT]
 class HttpResponse<T, RT> {
   final T data;
+
   final ResponseValue<RT> response;
   HttpResponse(this.data, this.response);
 }
 
-/// Pure value of http response
+/// real response data
+///
+/// - [data] any type
+/// - [statusCode] Response status code
 class ResponseValue<RT> {
   final RT? data;
   final int? statusCode;