+++ lib/src/http_response/http_response_extension.dart
@@ -1,0 +1,5 @@
+/// Covert response to dart class or something else
+typedef ResponseConverter<R, RT> = R Function(RT map);
+
+/// [dio.Response] convertor
+typedef DioConverter<R> = ResponseConverter<R, Map<String, dynamic>>;