+++ test/error_handler_test.dart
@@ -1,17 +1,59 @@
+import 'package:dio/dio.dart';
 import 'package:error_handler/error_handler.dart';
-import 'package:test/test.dart';
+import 'package:http_mock_adapter/http_mock_adapter.dart';
+import 'package:test/expect.dart';
+import 'package:test/scaffolding.dart';
 
-import 'api_call.dart';
+import 'post_test.dart';
 
+/// first create [Dio] api call
+FutureResponse<Post> getPost() async {
+  final dio = Dio(BaseOptions());
+  final dioAdapter = DioAdapter(dio: dio);
+  const path = "https://theKeySoftware.com/";
+  final data = {
+    "id": 1,
+    "userId": 1,
+    "title": "title",
+    "decryption": "decryption"
+  };
+  dioAdapter.onGet(
+    path,
+    (server) => server.reply(
+      200,
+      data,
+      delay: const Duration(seconds: 1),
+    ),
+  );
+  final response = await dio.get(path);
+
+  return response.convert(Post.fromJson);
+}
+
+/// wrap the api call with [safeApiCall]
 void main() {
-  setUp(() {
-    // Additional setup goes here.
+  group("ErrorHandler.stream", () {
+    test("stream", () {
+      errorHandler.stream(getPost).listen((event) {
+        event.whenOrNull(
+          data: (post, response) {
+            print("title");
+            expect(post.title, "title");
+          },
+        );
+      });
+    });
   });
-  group('Errors', () {
-    test('First Test', () async {
-      await safeApiCall(() => getUser({"name": "Alex"}))
-          .listen((event) => print(event))
-          .asFuture();
+  group("ErrorHandler.future", () {
+    test("future", () async {
+      final state = await errorHandler.future(getPost);
+
+      state.whenOrNull(
+        data: (post, response) {
+          print("title");
+          expect(post.title, "title");
+        },
+      );
     });
   });
 }