+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.4
+
+- add base class ErrorHandler
+- support both stream returning and future too
+- 
+- full support to [chopper](https://pub.dev/packages/chopper) client
 ## 1.0.3-dev
 
 - add [chopper](https://pub.dev/packages/chopper) support